BACKEND/requirements.txt
fastapi==0.100.0
uvicorn==0.23.1
pydantic==2.1.1
openai==1.6.0
python-dotenv==1.0.0
requests==2.32.3
PyYAML==6.0.2
